+++ js/constants.js
@@ -1,0 +1,23 @@
+const MESSAGES_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const NAMES_USERS = [
+  'Артём',
+  'Петя',
+  'Ольга',
+  'Светлана'
+];
+const MAX_ITEM_ID = 25;
+const IMG_FOLDER = 'photos/';
+const IMG_EXTEND = '.jpg';
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const AVATAR_FOLDER = 'img/avatar-';
+const AVATAR_EXTEND = '.svg';
+
+export {MESSAGES_COMMENTS, NAMES_USERS, MAX_ITEM_ID, IMG_FOLDER, IMG_EXTEND, MIN_LIKES, MAX_LIKES, AVATAR_FOLDER, AVATAR_EXTEND};
